Move actions from tests to lib
